requirements.txt
attrs==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
automat==22.10.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.0 ; python_version >= "3.11" and python_version < "4.0" and (platform_python_implementation != "PyPy" or os_name == "nt") and (platform_python_implementation != "PyPy" or implementation_name != "pypy")
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
constantly==23.10.4 ; python_version >= "3.11" and python_version < "4.0"
cryptography==43.0.0 ; python_version >= "3.11" and python_version < "4.0"
cssselect==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.15.4 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
hyperlink==21.0.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
incremental==24.7.2 ; python_version >= "3.11" and python_version < "4.0"
itemadapter==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
itemloaders==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
lxml==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.0.1 ; python_version >= "3.12" and python_version < "4.0" or python_version == "3.11"
outcome==1.3.0.post0 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
parsel==1.9.1 ; python_version >= "3.11" and python_version < "4.0"
protego==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and (platform_python_implementation != "PyPy" or os_name == "nt") and (platform_python_implementation != "PyPy" or implementation_name != "pypy")
pydispatcher==2.0.7 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation == "CPython"
pyodbc==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
pyopenssl==24.2.1 ; python_version >= "3.11" and python_version < "4.0"
pypydispatcher==2.1.2 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation == "PyPy"
pysocks==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
queuelib==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
requests-file==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
scrapy==2.11.2 ; python_version >= "3.11" and python_version < "4.0"
selenium==4.27.1 ; python_version >= "3.11" and python_version < "4.0"
service-identity==24.1.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==72.1.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.11" and python_version < "4.0"
tldextract==5.1.2 ; python_version >= "3.11" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.11" and python_version < "4.0"
trio==0.28.0 ; python_version >= "3.11" and python_version < "4.0"
twisted==24.7.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3[socks]==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
w3lib==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
websocket-client==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
zope-interface==7.0.1 ; python_version >= "3.11" and python_version < "4.0"
